Refactoring and bugfix
